Factor out plots
